backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cassandra-yt-mcp-backend"
version = "0.1.0"
description = "Private backend API for Cassandra YT MCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "pydantic>=2.10.0",
  "python-dotenv>=1.0.0",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.9.0",
]
gpu = [
  "nemo_toolkit[asr]>=2.0.0",
  "pyannote.audio>=3.3.0",
  "torch>=2.2.0",
  "torchaudio>=2.2.0",
]

[project.scripts]
cassandra-yt-mcp-backend = "cassandra_yt_mcp.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
